Fix buttons
